Release 1.9.0

+++ tests/test_router_ospf_vrf.rb
@@ -300,6 +300,8 @@ class TestRouterOspfVrf < CiscoTestCase
     assert_equal(metric, vrf.default_metric,
                  "Error: #{vrf.name} vrf, default-metric get value mismatch")
 
+    skip_versions = ['7.0.3.I7.3']
+    return unless step_unless_legacy_defect(skip_versions, 'CSCvi74876: Cannot remove default metric')
     vrf.default_metric = vrf.default_default_metric
     refute_match_vrf_line(vrf.parent.name, vrf.name, pattern)
     vrf.parent.destroy
